Use more static_cast explicit casts

+++ launcher/src/accountmanager.cpp
@@ -121,7 +121,7 @@ bool AccountManager::hasAnyAccounts() const
 
 int AccountManager::numAccounts() const
 {
-    return m_accounts.count();
+    return static_cast<int>(m_accounts.count());
 }
 
 #include "moc_accountmanager.cpp"
